ai-orchestrator-2026/requirements.txt
# === AI Orchestrator 2026 — Зависимости ===

# LLM клиент с поддержкой множества провайдеров
litellm>=1.40.0
httpx>=0.27.0

# Валидация данных
pydantic>=2.5.0

# Асинхронная SQLite для версионирования
aiosqlite>=0.19.0

# Переменные окружения
python-dotenv>=1.0.0

# Наблюдаемость (OpenTelemetry)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# ChromaDB — персистентное векторное хранилище
chromadb>=0.5.0
sentence-transformers>=3.0.0

# DSPy — автоматическая оптимизация промптов (MIPROv2)
dspy-ai>=2.4.0

# PyTorch (зависимость sentence-transformers)
torch>=2.0.0
